Fix to bug regarding int32 overflow (#660)
* Fix to bug regarding int32 overflow

* The peer response timer and connection activity timer return
`uint64_t` values in microseconds. All related logic, including
the non-responding member checker, should also use `uint64_t`,
even when converting the values to milliseconds.

* Otherwise, a 32-bit integer overflow could cause incorrect
behavior, such as including a non-responding member in the quorum.
The overflow occurs at `2^31` milliseconds, or approximately 24.8
days.

* Update PR

+++ include/libnuraft/raft_server.hxx
@@ -1032,17 +1032,18 @@ protected:
     int32 get_quorum_for_election();
     int32 get_quorum_for_commit();
     int32 get_leadership_expiry();
-    std::list<ptr<peer>> get_not_responding_peers(int expiry = 0);
-    size_t get_not_responding_peers_count(int expiry = 0, uint64_t required_log_idx = 0);
+    std::list<ptr<peer>> get_not_responding_peers(uint64_t expiry = 0);
+    size_t get_not_responding_peers_count(uint64_t expiry = 0,
+                                          uint64_t required_log_idx = 0);
     size_t get_num_stale_peers();
     static bool is_excluded_from_quorum(const peer& pp,
-                                        int32_t resp_elapsed_ms,
-                                        int32_t expiry,
+                                        uint64_t resp_elapsed_ms,
+                                        uint64_t expiry,
                                         uint64_t required_log_idx,
                                         bool include_self_mark_down = true);
 
     void for_each_voting_members(
-        const std::function<void(const ptr<peer>&, int32_t)>& callback);
+        const std::function<void(const ptr<peer>&, uint64_t)>& callback);
 
     ptr<resp_msg> handle_append_entries(req_msg& req);
     ptr<resp_msg> handle_prevote_req(req_msg& req);
